refactor(0.15.15): 命名預覽邏輯搬進 shared，供搜尋頁空狀態共用
T4 的搜尋頁空狀態要顯示「你的檔案會被整理成什麼樣」的範例，算法必須與設定頁
「預覽」那一列完全相同——不能在搜尋頁自己再寫一份，否則兩處遲早漂移。

把 stripFolderExcludedTokens / normalizeFolderLayers 從 pages/settings/chip-editor.js
搬到 shared/naming-preview.js（是搬移不是複製；那兩支是純命名邏輯，與 ChipEditor
膠囊 widget 無關），新增 buildNamingPreview 把 state-config.js 的 _previewWith
整段邏輯抽成純函式，_previewWith 改為委派。

對使用者零可見差異：設定頁命名預覽的輸出逐位元不變。

- 13 個既有測項（〔g〕–〔h5〕）跟著搬到 shared/__tests__/naming-preview.test.mjs，
  斷言與名稱逐字保留；新增 3 個測項覆蓋 buildNamingPreview 的三個分支
  （建立資料夾關閉、{suffix} 不進資料夾層、formatVariables 為空時的降級）

驗證：
- node --test naming-preview.test.mjs 16/16、chip-editor.test.mjs 7/7
- mutation 1/1 PASS；另加兩條 card 沒宣稱的獨立 mutation：
  slice(0,3)→slice(0,4) 讓〔h〕〔h2〕〔h5〕轉紅、folder_ok===false→folder_ok 讓〔i2〕轉紅
- ./scripts/check.sh --fast 全綠

Task: TASK-146a-T2（實作 gemini-3.8-flash-high，review grok-4.5）
